Move appender over to PyTestStrategy
